refactor: renaming and restructure
