+++ src/main/java/com/chess/backend/restController/service/GetChessboardService.java
@@ -1,13 +1,13 @@
 package com.chess.backend.restController.service;
 
-import com.chess.backend.GameController;
+import com.chess.backend.service.GameService;
 import com.chess.backend.restController.objects.ChessboardObject;
 import org.springframework.stereotype.Service;
 
 @Service
 public class GetChessboardService {
     public ChessboardObject getChessboard(int gameID){
-        GameController gc = GameController.getInstance();
+        GameService gc = GameService.getInstance();
 
         return new ChessboardObject(gameID, gc.getChessboard(gameID));
     }
