+++ src/main/java/com/chess/backend/restController/service/MoveRequestService.java
@@ -1,6 +1,6 @@
 package com.chess.backend.restController.service;
 
-import com.chess.backend.GameController;
+import com.chess.backend.service.GameService;
 import com.chess.backend.restController.objects.MoveRequestObject;
 import org.springframework.stereotype.Service;
 
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 public class MoveRequestService {
     public MoveRequestObject getPossibleMoves(int gameID, String pieceID, int[] piecePosition){
 
-        GameController gc = GameController.getInstance();
+        GameService gc = GameService.getInstance();
 
         return new MoveRequestObject(gameID, pieceID, piecePosition, gc.getPossibleMoves(gameID, pieceID, piecePosition));
     }
