+++ src/main/java/com/chess/backend/restController/service/NewGameService.java
@@ -1,12 +1,12 @@
 package com.chess.backend.restController.service;
 
-import com.chess.backend.GameController;
+import com.chess.backend.service.GameService;
 import org.springframework.stereotype.Service;
 
 @Service
 public class NewGameService {
     public int getNewGameID(String[] players){
-        GameController gc = GameController.getInstance();
+        GameService gc = GameService.getInstance();
 
         if(gc.createNewGame(players)){
             return gc.getGameID();
