+++ src/main/java/com/chess/backend/restController/service/VerifyExecutedMoveService.java
@@ -1,13 +1,13 @@
 package com.chess.backend.restController.service;
 
-import com.chess.backend.GameController;
+import com.chess.backend.service.GameService;
 import com.chess.backend.restController.objects.VerifyExecutedMoveObject;
 import org.springframework.stereotype.Service;
 
 @Service
 public class VerifyExecutedMoveService {
     public VerifyExecutedMoveObject verifyExecutedMove(int gameID, String pieceID, int[] previousPiecePosition, int[] newPiecePosition){
-        GameController gc = GameController.getInstance();
+        GameService gc = GameService.getInstance();
 
         boolean isMoveValid = gc.verifyExecutedMove(gameID, pieceID, previousPiecePosition, newPiecePosition);
         Object[][] chessboard = gc.getChessboard(gameID);
